+++ lib/cucumber/glue/dsl.rb
@@ -107,14 +107,6 @@ module Cucumber
         value.nil? ? default : value
       end
 
-      # Registers a proc that will run after Cucumber is configured. You can register as
-      # as you want (typically from ruby scripts under <tt>support/hooks.rb</tt>).
-      #
-      # DEPRECATED: please use InstallPlugin or BeforeAll instead
-      def AfterConfiguration(&proc)
-        Dsl.register_rb_hook('after_configuration', [], proc)
-      end
-
       # Registers a proc that will run after Cucumber is configured in order to install an external plugin.
       def InstallPlugin(&proc)
         Dsl.register_rb_hook('install_plugin', [], proc)