+++ lib/cucumber/runtime.rb
@@ -72,7 +72,6 @@ module Cucumber
       )
 
       load_step_definitions
-      fire_after_configuration_hook
       fire_install_plugin_hook
       fire_before_all_hook unless dry_run?
       # TODO: can we remove this state?
@@ -112,10 +111,6 @@ module Cucumber
     end
 
     private
-
-    def fire_after_configuration_hook # :nodoc:
-      @support_code.fire_hook(:after_configuration, @configuration)
-    end
 
     def fire_install_plugin_hook # :nodoc:
       @support_code.fire_hook(:install_plugin, @configuration, registry_wrapper)